# Optimize size-of by capturing all types by kind