```
name         old time/op    new time/op    delta
WriteCSV-44    6.19ms ± 2%    5.52ms ± 3%  -10.82%  (p=0.000 n=20+20)

name         old alloc/op   new alloc/op   delta
WriteCSV-44    5.44MB ± 0%    5.44MB ± 0%     ~     (p=0.755 n=19+20)

name         old allocs/op  new allocs/op  delta
WriteCSV-44     61.5k ± 0%     61.5k ± 0%     ~     (all equal)
```